Add path and modified time args to `build_parquet` (#17)
The `incremental` mode for `build_parquet` looks for path and modified
time columns in the table to determine if a file is new/changed. Since
these column names can vary, add arguments so the user can specify.

TODO: it is a bit clumsy to have these pass-through arguments. There
might be a better way to refactor.